Change signature of IClip::SetCacheHints to match Avisynth. Instead, Use dynamic_cast to detect cache clips.

+++ avxsynth/builtinfunctions/src/filters/combine.h
@@ -60,7 +60,7 @@ public:
     { return vi; }
   inline bool __stdcall GetParity(int n) 
     { return child1->GetParity(n); }
-  void __stdcall SetCacheHints(int cachehints,size_t frame_range) { };
+  void __stdcall SetCacheHints(int cachehints,int frame_range) { };
 
   static AVSValue __cdecl Create(AVSValue args, void*, IScriptEnvironment* env);
 
@@ -88,7 +88,7 @@ public:
   inline bool __stdcall GetParity(int n) 
     { return child1->GetParity(n); }
 
-  void __stdcall SetCacheHints(int cachehints,size_t frame_range) { };
+  void __stdcall SetCacheHints(int cachehints,int frame_range) { };
   static AVSValue __cdecl Create(AVSValue args, void*, IScriptEnvironment* env);
 
 private:
@@ -115,7 +115,7 @@ public:
     { return vi; }
   inline bool __stdcall GetParity(int n) 
     { return child[0]->GetParity(n); }
-  void __stdcall SetCacheHints(int cachehints,size_t frame_range) { };
+  void __stdcall SetCacheHints(int cachehints,int frame_range) { };
 
   static AVSValue __cdecl Create(AVSValue args, void*, IScriptEnvironment* env);
 
@@ -152,7 +152,7 @@ public:
     { return cache[0]->GetVideoInfo(); }
   bool __stdcall GetParity(int n);
 
-  void __stdcall SetCacheHints(int cachehints,size_t frame_range) { };
+  void __stdcall SetCacheHints(int cachehints,int frame_range) { };
 
   static AVSValue __cdecl Create(AVSValue args, void*, IScriptEnvironment* env);
   static AVSValue __cdecl Create_Range(AVSValue args, void*, IScriptEnvironment* env);
