+++ avxsynth/builtinfunctions/src/filters/layer.h
@@ -69,7 +69,7 @@ public:
     { return child1->GetParity(n); }
 
   static AVSValue __cdecl Create(AVSValue args, void*, IScriptEnvironment* env);
-  void __stdcall SetCacheHints(int cachehints,size_t frame_range) { };
+  void __stdcall SetCacheHints(int cachehints,int frame_range) { };
 
 private:
   const PClip child1, child2;
@@ -186,7 +186,7 @@ public:
     { return vi; }
   inline virtual bool __stdcall GetParity(int n) 
     { return child1->GetParity(n); }
-  void __stdcall SetCacheHints(int cachehints,size_t frame_range) { };
+  void __stdcall SetCacheHints(int cachehints,int frame_range) { };
 
   static AVSValue __cdecl Create(AVSValue args, void*, IScriptEnvironment* env);
 
@@ -218,7 +218,7 @@ public:
     { return vi; }
   inline virtual bool __stdcall GetParity(int n) 
     { return child1->GetParity(n); }
-  void __stdcall SetCacheHints(int cachehints,size_t frame_range) { };
+  void __stdcall SetCacheHints(int cachehints,int frame_range) { };
 
   static AVSValue __cdecl Create(AVSValue args, void*, IScriptEnvironment* env);
