+++ avxsynth/common/include/core/cache.h
@@ -56,7 +56,7 @@ public:
   Cache(PClip _child, IScriptEnvironment* env);
   ~Cache();
   PVideoFrame __stdcall GetFrame(int n, IScriptEnvironment* env);
-  void __stdcall SetCacheHints(int cachehints,size_t frame_range);
+  void __stdcall SetCacheHints(int cachehints,int frame_range);
   static AVSValue __cdecl Create_Cache(AVSValue args, void*, IScriptEnvironment* env);
   void __stdcall GetAudio(void* buf, __int64 start, __int64 count, IScriptEnvironment* env);
